Make composite non-default experimental feature.
./configure --with-composite - to enable compositing features.

Let's face it. The reasons are:
 - Tray doesn't work with composite mode.
 - Composite does true transparency, but xshape isn't here, so, this
   feature is useless. A good screenshot for making impressions on
   your friends can't be a reason.
 - Composite requires compositing manager.

And personally I think that it's not the age of compositing yet.
Maybe it's about eye-candy. But after rethink of my philosophy for
bmpanel, it's definitely about usability and KISS.

Panel should be pretty enough and usable enough. Running on bleeding
edge was my mistake. Also I have to say, that compositing probably
will be removed in future versions.